[SPE-1129] Update logic to check if the cert has been renewed
